python_service/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pdfplumber
openai
python-dotenv
pillow
chromadb
sentence-transformers
rank-bm25
jieba
torch
python-docx
